requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
factory-boy==3.2.1
Faker==9.8.1
idna==3.3
ipython==7.29.0
jedi==0.18.0
matplotlib-inline==0.1.3
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
